Move XNA classes into XNA namespace, PacketTypes into TerrariaApi.Server, and remove TileCommand
