+++ Terraria/Player.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using System.IO;
 using System.Security.Cryptography;
 using System.Text;
+using XNA;
+
 namespace Terraria
 {
 	public class Player
