+++ TerrariaApi.Server/EventArgs/ChatReceivedEventArgs.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using XNA;
 
 namespace TerrariaApi.Server
 {
