+++ TerrariaApi.Server/HookManager.cs
@@ -1,12 +1,12 @@
 ﻿using System;
 using System.ComponentModel;
 using System.Diagnostics;
-using System.Text;
+using System.IO;
 using System.Linq;
 using System.Security.Cryptography;
-
+using System.Text;
 using Terraria;
-using System.IO;
+using XNA;
 
 namespace TerrariaApi.Server
 {
