--- TileCommand.cs
@@ -1,9 +1,0 @@
-using System;
-public enum TileCommand
-{
-	KillTile,
-	PlaceTile,
-	KillWall,
-	PlaceWall,
-	KillTileNoItem
-}
