+++ XNA/IPackedVector.cs
@@ -1,0 +1,18 @@
+using System;
+
+namespace XNA
+{
+    public interface IPackedVector
+    {
+        void PackFromVector4(Vector4 vector);
+        Vector4 ToVector4();
+    }
+    public interface IPackedVector<TPacked> : IPackedVector
+    {
+        TPacked PackedValue
+        {
+            get;
+            set;
+        }
+    }
+}
